requirements.txt
# ──────────────────────────────────────────────
# FinSage — Personal Finance Agentic AI
# requirements.txt  (Chunk 20 — final)
# ──────────────────────────────────────────────
# Install:  pip install -r requirements.txt
# Python  : 3.10 or 3.11 recommended
# ──────────────────────────────────────────────

# ── LLM & Agent Framework ─────────────────────
langchain>=0.2.16,<0.3.0
langchain-core>=0.2.38,<0.3.0
langchain-groq>=0.1.9,<0.2.0
langgraph>=0.2.0,<0.3.0

# ── Vector Store & Embeddings ─────────────────
chromadb>=0.5.3,<0.6.0
sentence-transformers>=3.0.1,<4.0.0

# ── Streamlit UI ──────────────────────────────
streamlit>=1.37.0,<2.0.0

# ── Evaluation ────────────────────────────────
# RAGAS requires datasets + specific langchain pins
ragas>=0.1.14,<0.2.0
datasets>=2.20.0

# ── Environment & Config ──────────────────────
python-dotenv>=1.0.1,<2.0.0

# ── Utilities ─────────────────────────────────
# numpy is pulled in by sentence-transformers; pin to avoid conflicts
numpy>=1.26.0,<2.0.0

# ── Testing ───────────────────────────────────
# Standard library unittest is used (no extra install needed).
# Uncomment below if you prefer pytest:
# pytest>=8.0.0
